+++ src/components/index.js
@@ -3,3 +3,4 @@ export * from './SearchTextareaField';
 export * from './MultiSelectionFacet';
 export * from './FacetOptionFormatter';
 export * from './SearchFilters';
+export * from './MarcAuthoritiesAppContext';