+++ src/index.js
@@ -2,17 +2,22 @@ import 'core-js/stable';
 import 'regenerator-runtime/runtime';
 
 import PropTypes from 'prop-types';
+
 import {
   Switch,
 } from 'react-router-dom';
 
-import { Route } from '@folio/stripes/core';
+import {
+  Route,
+} from '@folio/stripes/core';
 
 import {
   SearchRoute,
   AuthorityViewRoute,
   AuthorityQuickMarcEditRoute,
 } from './routes';
+
+import { MarcAuthoritiesAppContext } from './components';
 
 const propTypes = {
   match: PropTypes.object.isRequired,
@@ -27,15 +32,18 @@ const MarcAuthorities = ({
   },
 }) => {
   return (
-    <Switch>
-      <Route path={`${path}/quick-marc`} component={AuthorityQuickMarcEditRoute} />
-      <Route
-        path={path}
-        component={SearchRoute}
-      >
-        <Route path={`${path}/authorities/:id`} component={AuthorityViewRoute} />
-      </Route>
-    </Switch>
+    <>
+      <MarcAuthoritiesAppContext />
+      <Switch>
+        <Route path={`${path}/quick-marc`} component={AuthorityQuickMarcEditRoute} />
+        <Route
+          path={path}
+          component={SearchRoute}
+        >
+          <Route path={`${path}/authorities/:id`} component={AuthorityViewRoute} />
+        </Route>
+      </Switch>
+    </>
   );
 };